requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-mysql==4.8.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==1.1.1
install==1.3.5
mysqlclient==2.1.1
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
